agragando hasta politicas de acceso

+++ app/Http/Controllers/Admin/PostsController.php
@@ -14,89 +14,52 @@ class PostsController extends Controller
 {
     public function index()
     {
-    	$posts = Post::all();
+    	// $posts = Post::where('user_id',auth()->id())->get();
+        $posts = auth()->user()->posts;
     	return view('admin.posts.index',compact('posts'));
     }
 
-    // public function create()
-    // {
-    // 	$tags = Tag::all();
-    // 	$categorias = Category::all();
-    // 	return view('admin.posts.create',compact('categorias','tags'));
-    // }
-
     public function store(Request $request)
     {
+        
+        $this->authorize('create', new Post);
+
         $this->validate($request,[
             'title' => 'required|min:3',
         ]);
 
-        $post = Post::create( $request->only('title') );
-
-        
+        // $post = Post::create( $request->only('title') );
+        $post = Post::create( $request->all() );
 
         return redirect()->route('admin.posts.edit',$post);
     }
 
     public function edit(Post $post)
     {
-        $tags = Tag::all();
-        $categories = Category::all();
-        return view('admin.posts.edit',compact('categories','tags','post'));
+        $this->authorize('view',$post);
+
+        return view('admin.posts.edit',[
+            'post' => $post,
+            'tags' => Tag::all(),
+            'categories' => Category::all()
+        ]);
     }
 
     public function update(Post $post,StorePostRequest $request)
     {
-        $post->update($request->all());
+        
+        $this->authorize('update',$post);
 
+        $post->update($request->all());
         $post->syncTags($request->get('tags'));
         
         return redirect()->route('admin.posts.edit',$post)->with('flash','tu publicación ha sido guardada');
-
-        // dd($request->all());
-        // dd($request->all());
-        // $request->published_at = Carbon::parse($request->get('published_at'));
-        // // dd($request->published_at);
-        // Post::create($request->all());
-        // return $request;
-
-        // return $request->all();
-
-        // $post->title = $request->get('title');
-        // $post->body = $request->get('body');
-        // $post->iframe = $request->get('iframe');
-        // $post->excerpt = $request->get('excerpt');
-        // $post->published_at = $request->get('published_at');
-        // $post->category_id = $request->get('category_id');
-        // $post->save();
-
-        // $tags = collect($request->get('tags'))->map(function($tag){
-        //     return Tag::find($tag) ? $tag : Tag::create(['name' => $tag])->id;
-        // });
-
-        // $post->tags()->sync($tags);
-        // $tags = [];
-        // foreach ($request->get('tags') as $tag) {
-        //     $tags[] = Tag::find($tag)
-        //             ? $tag
-        //             : Tag::create(['name' => $tag])->id;
-        // }
-
-
 
     }
 
     public function destroy(Post $post)
     {
-
-        // $post->photos()->delete();
-        // foreach ($post->photos as $photo) {
-        //     $photo->delete();
-        // }
-        // $post->photos->each(function($photo){
-        //     $photo->delete();
-        // });
-
+        $this->authorize('delete',$post);
         $post->delete();
         return redirect()->route('admin.posts.index')->with('flash','tu publicación ha sido eliminada');
     }
